app(fix): move components dir out of assets
